restructure files and add to SelectMultipleFileTypes
